feat(glens): embeddable Lens surface (LensEmbed) for #1830 Tab II
Extract the docked-panel chat guts into <LensChat> (headless: messages
viewport + composer + SSE streaming loop). LensPanel becomes a thin
docked shell (aside + resize handle + header + Escape) that renders
LensChat inside.

Add <LensEmbed> — inline Lens capability for the #1830 embedded-workflow
demo on /theguard/try Tab II. No AppShell chrome, no dock, fills its
parent container:

  <LensEmbed sessionId={runThreadId} initialQuery="…" height={480} />

Attaches to an existing thread via sessionId (host page owns the run).
Reuses the same streaming, markdown parser, and AnswerBubble as the
other two surfaces — one loop, three renderings.

LensPanel: 244 → 16 line diff. Any future chat capability (retry, edit,
copy) lands in LensChat and hits all three surfaces at once.
